+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- new theme of BaseExpander
+
 ### Fixed
 
 - visibility of text and icon in disabled BaseButton in dark theme